apps/api/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
sqlalchemy==1.4.41
psycopg2-binary==2.9.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.24.1
redis==5.0.1
